Use aviator as a REST interface for keystone_tenant provider
This commit migrates the keystone_tenant type and provider to use aviator
instead of command-line utilities for authenticating to the API and
creating, editing, and destroying tenants. It uses the aviator parent
provider from puppet-openstacklib in order to authenticate against the
REST API and make web requests. The keystone_tenant type uses the
openstacklib aviator utility to add new parameters to allow
authentication credentials to be passed in.

Change-Id: I73d0083e5357fd88c76d00f9f1156f8615844c80
Implements: blueprint use-aviator-in-module-resources